feat: organize files
