+++ src/components/sections/Features.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useLanguage } from "../contexts/LanguageContext";
+import { useLanguage } from "../../contexts/LanguageContext";
 
 export const Features = ({ data }) => {
   const { language } = useLanguage();
