+++ src/components/sections/Services.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useLanguage } from "../contexts/LanguageContext";
+import { useLanguage } from "../../contexts/LanguageContext";
 
 export const Services = ({ data }) => {
   const { language } = useLanguage();
