+++ src/components/sections/Team.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useLanguage } from "../contexts/LanguageContext";
+import { useLanguage } from "../../contexts/LanguageContext";
 
 export const Team = ({ data, description }) => {
   const { language } = useLanguage();
